// src/app/launch-module/models/page.enum.ts
export enum PageModel {
    PAGE_HOME = 'PAGE_HOME',
    PAGE_RESEARCH = 'PAGE_RESEARCH',
    PAGE_MOVIE = 'PAGE_MOVIE',
    PAGE_SERIES = 'PAGE_SERIES',
    PAGE_CATALOG = 'PAGE_CATALOG',
    PAGE_MYLIST = 'PAGE_MYLIST',
    PAGE_EDITION = 'PAGE_EDITION',
    PAGE_USER = 'PAGE_USER',
    PAGE_LICENSE = 'PAGE_LICENSE',
    DEFAULT_PAGE = 'DEFAULT_PAGE'
}